stubs: Improve enumfield type stubs
* Add EnumIntegerfield and IntEnum.

  - This will be needed for following commits.

* Move class definitions to correct submodules.
