+++ stubs/enumfields/enums.pyi
@@ -1,0 +1,9 @@
+import enum
+from typing import Sequence, Tuple
+
+class Enum(enum.Enum):
+    @classmethod
+    def choices(cls) -> Sequence[Tuple[str, str]]: ...
+    def __str__(self) -> str: ...
+
+class IntEnum(int, Enum): ...
